pfeiffer_tc400: Move pfeiffer tc400 agent into package
