--- agents/pfeiffer_tc400/requirements.txt
@@ -1,1 +1,0 @@
-pfeiffer-vacuum-protocol==0.4
