+++ docker-compose.yml
@@ -46,15 +46,6 @@ services:
       - "socs"
 
   # --------------------------------------------------------------------------
-  # Pfeiffer Turbo Controller
-  # --------------------------------------------------------------------------
-  ocs-pfeiffer-tc400-agent:
-    image: "ocs-pfeiffer-tc400-agent"
-    build: ./agents/pfeiffer_tc400/
-    depends_on:
-      - "socs"
-
-  # --------------------------------------------------------------------------
   # CHWP picoscope agent
   # --------------------------------------------------------------------------
   ocs-hwp-picoscope-agent:
