+++ setup.py
@@ -40,6 +40,7 @@ setup(name='socs',
           'sqlalchemy',
           'pysnmp',
           'requests',
+          'pfeiffer-vacuum-protocol==0.4',
           'xy_stage_control @ git+https://github.com/kmharrington/xy_stage_control.git@main',
           'soaculib @ git+https://github.com/simonsobs/soaculib.git@master',
       ],
