+++ tests/integration/test_pfeiffer_tc400_agent_integration.py
@@ -49,7 +49,7 @@ def format_reply(data):
 
 wait_for_crossbar = create_crossbar_fixture()
 run_agent = create_agent_runner_fixture(
-    '../agents/pfeiffer_tc400/pfeiffer_tc400_agent.py', 'tc400_agent')
+    '../socs/agents/pfeiffer_tc400/agent.py', 'tc400_agent')
 client = create_client_fixture('pfeifferturboA')
 emulator = create_device_emulator({}, relay_type='tcp')
